+++ lib/dlms-server/src/server/server_dispatch_trace.cpp
@@ -1,0 +1,9 @@
+#include "dlms/server/server_dispatch_trace.hpp"
+
+namespace dlms {
+namespace server {
+
+IServerDispatchTraceSink::~IServerDispatchTraceSink() = default;
+
+} // namespace server
+} // namespace dlms